// src/enums/error-code.enum.ts
export enum ErrorCode {
    UNDEFINED = 10000,
    HEALTH_CHECK = 10001,
    INTERNAL_SERVER_ERROR = 10010,
    VALIDATION = 10020,
    ORM_VALIDATION = 10030,
    COMPANY_EXISTS = 10040,
    ENTRY_EXISTS = 10050,
    JWT_EMPTY_ENTRY = 10060,
    JWT_EMPTY_SITE_ADMIN = 10070,
    JWT_EMPTY_ACCOUNT = 10080,
    JWT_EMPTY_PAYLOAD = 10090,
    JWT_EMPTY_PAYLOAD_DATA = 10100,
    JWT_EMPTY_AUTHORIZATION = 10110,
    JWT_INVALID_SIGNATURE = 10120,
    JWT_MISSING_USER = 10130,
    JWT_MISSING_ROLES = 10140,
    JWT_MISSING_ACCOUNT = 10150,
    JWT_MISSING_SITE_ADMIN = 10160,
    ENTITY_NOT_FOUND = 10170,
    BAD_CODE = 10180,
    BAD_ACCESS_TOKEN = 10190,
    CONFLICT_SELF = 10200,
    CONFLICT_PERSONNEL_OK = 10210,
    VERSION_NOT_FOUND = 10220,
    CONFLICT_DEPENDENCY = 10230,
    AGENT_MISSING_BODY = 10240,
    AGENT_MISSING_FIELDS = 10250,
    AGENT_EMPTY_ACCOUNT = 10260,
    AGENT_ALERT = 10265,
    ACCOUNT_NOT_FOUND = 10270,
    ACCOUNT_PENDING = 10271,
    ACCOUNT_MAINTENANCE = 10272,
    ACCOUNT_NON_PAYMENT = 10273,
    ACCOUNT_ADMIN_DISABLED = 10274,
    ACCOUNT_USER_DELETED = 10275,
    ACCOUNT_UNKNOWN_STATUS = 10276,
    CONNECTION_NOT_FOUND = 10280,
    USER_NOT_FOUND = 10290,
    USER_ROLES_NOT_FOUND = 10300,
    PERSONNEL_NOT_FOUND = 10310,
    PERSONNEL_NOT_ACTIVE = 10320,
    TOKEN_EXPIRED = 10330,
    DRATA_TERMS_NOT_AGREED_TO = 10340,
    BACKGROUND_CHECK_ALREADY_STARTED = 10350,
    BACKGROUND_CHECK_MISSING_PACKAGES = 10360,
    BACKGROUND_CHECK_UNAUTHORIZED = 10370,
    CONFLICT_MONITOR_CHECK_STATUS = 10380,
    CONFLICT_USER_IDENTITY_VERSION_CONTROL_LINK = 10390,
    CONFLICT_USER_INFRASTRUCTURE_CONTROL_LINK = 10400,
    PUBLISH_FRAMEWORK_MISSING_ELEMENT = 10410,
    CONFLICT_DUPLICATE_PROVIDER_CONNECTION = 10420,
    CONFLICT_MISMATCHED_EMAIL_DOMAIN = 10430,
    CONFLICT_COMMENT_AUTHOR = 10440,
    NONEXISTENT_COMMENT_THREAD = 10450,
    CONFLICT_COMMENT_EXISTS = 10460,
    SECURITY_REPORT_NOT_SHAREABLE = 10470,
    SECURITY_REPORT_SHARE_TOKEN_MISMATCH = 10480,
    RIPPLING_SERVICE_ERROR = 10490,
    ONBOARDING_STEP_LOCKED = 10500,
    ONBOARDING_NOT_COMPLETE = 10510,
    POLICY_NAME_UNAVAILABLE = 10520,
    MERGEDEV_LINK_TOKEN_FAILED = 10530,
    MERGEDEV_ACCESS_TOKEN_FAILED = 10540,
    REFRESH_TOKEN_NOT_FOUND = 10550,
    MAGIC_TOKEN_NOT_FOUND = 10555,
    POLICY_VERSION_TOO_LOW = 10560,
    POLICY_ALREADY_APPROVED = 10570,
    NO_ENABLED_FRAMEWORKS_FOR_POLICY = 10575,
    JWT_EMPTY_AUDITOR = 10580,
    JWT_MISSING_AUDITOR = 10590,
    AUDITOR_CLIENT_FORBIDDEN = 10600,
    JWT_AUDITOR_MISSING_CLIENT = 10610,
    CONFLICT_DUPLICATE_CLIENT_ID_CONNECTION = 10620,
    CONFLICT_CONNECTION_CLIENTID = 10630,
    CONFLICT_CONNECTION_AUTOPOPULATED_CLIENTID = 10640,
    AUTHENTICATION_ERROR_WORK_OS = 10700,
    AUTHENTICATION_ERROR_SOCKET = 10710,
    AUTHENTICATION_ERROR_WORK_OS_MISSING_PROFILE = 10720,
    AUTHENTICATION_ERROR_WORK_OS_MISSING_EMAIL = 10730,
    PAYMENT_REQUIRED = 10740,
    NO_CONTROLS_FOUND = 10750,
    CONFLICT_CONTROL_NEEDS_REQUIREMENT = 10760,
    INVALID_DOMAIN = 10770,
    CONFLICT_DUPLICATED_AUDITOR = 10780,
    CONFLICT_MISMATCH_DATA = 10790,
    EMPLOYEE_AUDITOR_CONFLICT = 10800,
    API_KEY_NOT_FOUND = 10810,
    CUSTOMER_UNAUTHORIZED = 10820,
    INVALID_CODE_PREFIX = 10900,
    DUPLICATE_CONTROL_CODE = 10910,
    REQUIREMENT_DOES_NOT_EXIST = 10920,
    SSO_ENFORCED = 10930,
    IDP_ENFORCED = 10940,
    SETTING_NOT_FOUND = 10950,
    CANNOT_CHANGE_DCF_CONTROL_CODE = 10960,
}